DC Elevator Feedback Controller
The purpose of this project was to design and simulate a closed-loop control system with feedback control. The elevator design had a micro-controller that controlled the movement of the actuator, which was an embedded C application that I wrote which provided more robust control of the system. 
These files attached in this respiratory include the following:
* word document of the final report
* .c files (containing the c code used for the STM Student Discovery Board Micro-controller)
* .png files (pictures of the system)
* note that MATLAB files were not included, and can be provided upon request.